requirements-full.txt
mcp>=1.0.0
requests>=2.28.0
fastf1>=3.0.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
websockets>=13.0
aiohttp>=3.9.0
certifi
pyjwt>=2.0.0
